requirements.txt
pygame
pillow
PySimpleGUI